+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * FIX for Create Step Definition Snippets Generates Reqnroll Using Statements for SpecFlow Projects #6
 * Fix for GH7 - "Find step definitions usages command not visible for SpecFlow projects
+* FIX for GH11: Find Step Definitions command fails for first time (partial fix)
 
 # v2024.1.49 - 2024-02-08