+++ CHANGELOG.md
@@ -1,10 +1,10 @@
 # [vNext]
+* FIX for GH18 - The Define Steps Command does not abide by the Reqnroll.json configuration setting for the stepDefinitionSkeletonStyle (CucumberExpression or RegexExpression)
 * FEATURE: FindUnused Step Definitions Command - from within a binding class a new context menu command, "FindUsusedStepDefinitions", is available. 
 	* This will list any Step Definition methods that are not matched by one or more Feature steps in the current project.
 
 * FIX for Create Step Definition Snippets Generates Reqnroll Using Statements for SpecFlow Projects #6
 * Fix for GH7 - "Find step definitions usages command not visible for SpecFlow projects
-* FIX for GH11: Find Step Definitions command fails for first time (partial fix)
 
 # v2024.1.49 - 2024-02-08