+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 * FIX for Create Step Definition Snippets Generates Reqnroll Using Statements for SpecFlow Projects #6
 * Fix for GH7 - "Find step definitions usages command not visible for SpecFlow projects
+* Project template have been updated to the latest Reqnroll and other dependency versions
 
 # v2024.1.49 - 2024-02-08